ASCS
Automatic Sewage cleaning system 
In this project we proposed the concept of "AUTOMATIC SEWAGE CLEANING SYSTEM", which replace the manual work of cleaning sewage by humans. AUTOMATIC SEWAGE CLEANING SYSTEM (ASCS) proposed to overcome the real time problems. This system  is used for automatic cleaning of sewage. This proposed system uses an automatic drain cleaning system that lets fluid flow through it but removed large solid waste like bottles, plastic and accumulates are blocked in sewage. Labour cleaning sewage leads to a high risk of them catching infections or poisoning due to large amount of waste/chemical  in them .so for reducing work of  humans and analysing several problems Water is the basic  need for the existence of life on earth. . More than 70% of water is being wasted only for our daily needs, Now a day automation plays a vital role in industrial application. By having this  project the instant block in the sewage system is being avoided. we proposed our project ASCS.